// src/lib/sanity/sanity.types.ts
export type Person = {
	_id: string;
	_type: 'person';
	_createdAt: string;
	_updatedAt: string;
	_rev: string;
	name?: string;
	slug?: Slug;
	email?: string;
	profileImage?: {
		asset?: {
			_ref: string;
			_type: 'reference';
			_weak?: boolean;
			[internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
		};
		media?: unknown;
		hotspot?: SanityImageHotspot;
		crop?: SanityImageCrop;
		alt?: string;
		_type: 'image';
	};
	role?: string;
	bio?: string;
	fullBio?: Array<{
		children?: Array<{
			marks?: Array<string>;
			text?: string;
			_type: 'span';
			_key: string;
		}>;
		style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
		listItem?: 'bullet' | 'number';
		markDefs?: Array<{
			href?: string;
			_type: 'link';
			_key: string;
		}>;
		level?: number;
		_type: 'block';
		_key: string;
	}>;
	isTeamMember?: boolean;
	isResident?: boolean;
	house?: {
		_ref: string;
		_type: 'reference';
		_weak?: boolean;
		[internalGroqTypeReferenceTo]?: 'house';
	};
	startDate?: string;
	endDate?: string;
	socialLinks?: {
		twitter?: string;
		linkedin?: string;
		github?: string;
		website?: string;
	};
	skills?: Array<string>;
	company?: string;
	displayOnWebsite?: boolean;
	displayOrder?: number;
};
// ...
export type SanityImageCrop = {
	_type: 'sanity.imageCrop';
	top?: number;
	bottom?: number;
	left?: number;
	right?: number;
};

export type SanityImageHotspot = {
	_type: 'sanity.imageHotspot';
	x?: number;
	y?: number;
	height?: number;
	width?: number;
};
// ...
export type Slug = {
	_type: 'slug';
	current?: string;
	source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol